pyproject.toml
[tool.poetry]
name = "bookswebsite"
version = "0.1.0"
description = ""
authors = ["Alexandra <alexandra20050416@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
